Add Get-AzSshCertificate cmdlet
This moves around all of the cert acquisition to a cmdlet and utilizes
the RMProfileClient (which holds the top-level GetSSHCertificate
function) to get the SSH cert off of the jwk and kid.

This has one major hack: The AuthenticationFactory has a special
AuthenticateSSH function that is not present on the
IAuthenticationFactory. This is because that interface lives in a
separate repo/nuget and I did not want to update the whole stack to make
it work. So in the RMProfileClient I make an assumed cast of the
session's authentication factory to the AuthenticationFactory class so i
can call AuthenticateSSH on it.